Introduce a 'RSASSAAlgorithm' abstract base class which all RSA algorithms derive from.  Add RS256Algorithm
